+++ src/main/java/com/grupo5e/morapack/repository/ProductoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ProductoRepository extends JpaRepository<Producto, Long> {
-    List<Producto> findByPedidoId(Long pedidoId);
+public interface ProductoRepository extends JpaRepository<Producto, Integer> {
+    List<Producto> findByPedidoId(Integer pedidoId);
     List<Producto> findByEstado(EstadoProducto estado);
 }
